ai-service/requirements.txt
Django==5.2.13
djangorestframework==3.15.2
requests==2.32.3
pika==1.3.2
neo4j==5.23.1
